+++ src/main/java/io/getunleash/impactmetrics/Gauge.java
@@ -3,19 +3,19 @@ package io.getunleash.impactmetrics;
 import java.util.Map;
 
 interface Gauge {
-    void set(long value);
+    void set(double value);
 
-    void set(long value, Map<String, String> labels);
+    void set(double value, Map<String, String> labels);
 
     void inc();
 
-    void inc(long value);
+    void inc(double value);
 
-    void inc(long value, Map<String, String> labels);
+    void inc(double value, Map<String, String> labels);
 
     void dec();
 
-    void dec(long value);
+    void dec(double value);
 
-    void dec(long value, Map<String, String> labels);
+    void dec(double value, Map<String, String> labels);
 }
